reorganised test suite, added tests for CLI
