+++ test/test_compiler/binoptest.c
