+++ test/test_compiler/literaltest.c
