+++ tests
@@ -14,19 +14,25 @@ fi
 printf "Build successful\n\n"
 
 printf "====================\n"
-printf "Running C tests...\n"
+printf "Running Compiler tests...\n"
 ./YASLTEST
 C_TESTS_OUTPUT=$?
-printf "C tests exited with code $C_TESTS_OUTPUT\n\n"
+printf "Compiler tests exited with code $C_TESTS_OUTPUT\n\n"
 
 printf "====================\n"
-printf "Running Perl tests...\n"
+printf "Running VM tests...\n"
 ( cd test ; perl vmtest.pl )
-PERL_TESTS_OUTPUT=$?
-printf "Perl tests exited with code $PERL_TESTS_OUTPUT\n\n"
+VM_TESTS_OUTPUT=$?
+printf "VM tests exited with code $VM_TESTS_OUTPUT\n\n"
 
 printf "====================\n"
-[ $C_TESTS_OUTPUT$PERL_TESTS_OUTPUT = 00 ]
+printf "Running CLI tests...\n"
+( cd test ; perl clitest.pl )
+CLI_TESTS_OUTPUT=$?
+printf "CLI tests exited with code $CLI_TESTS_OUTPUT\n\n"
+
+printf "====================\n"
+[ $C_TESTS_OUTPUT$VM_TESTS_OUTPUT$CLI_TESTS_OUTPUT = 000 ]
 TESTS_EXIT=$?
 printf "Tests exited with code $TESTS_EXIT\n"
